+++ settings.gradle.kts
@@ -57,9 +57,9 @@ include(":pvptools")
 include(":reorderprayers")
 include(":runedoku")
 include(":socket")
-include(":socket-player-status")
-include(":socket-sotetseg")
-include(":socket-special-counter")
+include(":socketplayerstatus")
+include(":socketsotetseg")
+include(":socketspecialcounter")
 include(":specbar")
 include(":spellbook")
 include(":strongholdofsecurity")
