--- socket-player-status/src/main/java/net/runelite/client/plugins/playerstatus/gametimer/GameTimer.java
@@ -1,91 +1,0 @@
-/*
- * Copyright (c) 2017, Seth <Sethtroll3@gmail.com>
- * Copyright (c) 2017, Adam <Adam@sigterm.info>
- * Copyright (c) 2018, Jordan Atwood <jordan.atwood423@gmail.com>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package net.runelite.client.plugins.playerstatus.gametimer;
-
-import java.time.Duration;
-import java.time.temporal.ChronoUnit;
-import lombok.AccessLevel;
-import lombok.Getter;
-import net.runelite.api.GraphicID;
-import net.runelite.api.ItemID;
-import net.runelite.api.SpriteID;
-
-@Getter(AccessLevel.PUBLIC)
-public enum GameTimer
-{
-
-	VENGEANCE(SpriteID.SPELL_VENGEANCE, GameTimerImageType.SPRITE, "Vengeance", 30,
-		ChronoUnit.SECONDS),
-	OVERLOAD(ItemID.OVERLOAD_4, GameTimerImageType.ITEM, "Overload", 5, ChronoUnit.MINUTES, true),
-	OVERLOAD_RAID(ItemID.OVERLOAD_4_20996, GameTimerImageType.ITEM, "Overload", 5, ChronoUnit.MINUTES,
-		true),
-	PRAYER_ENHANCE(ItemID.PRAYER_ENHANCE_4, GameTimerImageType.ITEM, "Prayer enhance", 290,
-		ChronoUnit.SECONDS, true),
-	STAMINA(ItemID.STAMINA_POTION4, GameTimerImageType.ITEM, "Stamina", 2, ChronoUnit.MINUTES, true),
-	IMBUED_HEART(ItemID.IMBUED_HEART, GameTimerImageType.ITEM, "Imbued heart", GraphicID.IMBUED_HEART,
-		420, ChronoUnit.SECONDS);
-
-	private final Duration duration;
-	private final Integer graphicId;
-	private final String description;
-	private final boolean removedOnDeath;
-	private final Duration initialDelay;
-	private final int imageId;
-	private final GameTimerImageType imageType;
-
-	GameTimer(int imageId, GameTimerImageType idType, String description, Integer graphicId, long time, ChronoUnit unit, long delay, boolean removedOnDeath)
-	{
-		this.description = description;
-		this.graphicId = graphicId;
-		this.duration = Duration.of(time, unit);
-		this.imageId = imageId;
-		this.imageType = idType;
-		this.removedOnDeath = removedOnDeath;
-		this.initialDelay = Duration.of(delay, unit);
-	}
-
-	GameTimer(int imageId, GameTimerImageType idType, String description, Integer graphicId, long time, ChronoUnit unit, boolean removedOnDeath)
-	{
-		this(imageId, idType, description, graphicId, time, unit, 0, removedOnDeath);
-	}
-
-	GameTimer(int imageId, GameTimerImageType idType, String description, long time, ChronoUnit unit, boolean removeOnDeath)
-	{
-		this(imageId, idType, description, null, time, unit, removeOnDeath);
-	}
-
-	GameTimer(int imageId, GameTimerImageType idType, String description, long time, ChronoUnit unit)
-	{
-		this(imageId, idType, description, null, time, unit, false);
-	}
-
-	GameTimer(int imageId, GameTimerImageType idType, String description, Integer graphicId, long time, ChronoUnit unit)
-	{
-		this(imageId, idType, description, graphicId, time, unit, false);
-	}
-}
