--- socket-player-status/src/main/java/net/runelite/client/plugins/playerstatus/marker/AbstractMarker.java
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2020, Charles Xu <github.com/kthisiscvpv>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package net.runelite.client.plugins.playerstatus.marker;
-
-import java.awt.Graphics2D;
-import java.awt.image.BufferedImage;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-
-public abstract class AbstractMarker
-{
-
-	@Setter(AccessLevel.PUBLIC)
-	@Getter(AccessLevel.PUBLIC)
-	private BufferedImage baseImage;
-
-	public BufferedImage getImage(int size)
-	{
-		BufferedImage baseImage = this.getBaseImage();
-		if (baseImage == null)
-		{
-			return null;
-		}
-
-		double height = baseImage.getHeight() > 0 ? baseImage.getHeight() : 1;
-		double scale = ((double) size) / height;
-
-		int newWidth = (int) Math.ceil(scale * baseImage.getWidth());
-		BufferedImage realImage = new BufferedImage(newWidth, size, baseImage.getType());
-
-		Graphics2D g2d = realImage.createGraphics();
-		g2d.drawImage(baseImage, 0, 0, newWidth, size, null);
-		g2d.dispose();
-
-		return realImage;
-	}
-}
